+++ lib/active_shipping/carriers/fedex.rb
@@ -143,7 +143,7 @@ module ActiveShipping
     end
 
     def find_rates(origin, destination, packages, options = {})
-      options = @options.update(options)
+      options = @options.merge(options)
       packages = Array(packages)
 
       rate_request = build_rate_request(origin, destination, packages, options)
@@ -154,7 +154,7 @@ module ActiveShipping
     end
 
     def find_tracking_info(tracking_number, options = {})
-      options = @options.update(options)
+      options = @options.merge(options)
 
       tracking_request = build_tracking_request(tracking_number, options)
       xml = commit(save_request(tracking_request), (options[:test] || false))
@@ -164,7 +164,7 @@ module ActiveShipping
 
     # Get Shipping labels
     def create_shipment(origin, destination, packages, options = {})
-      options = @options.update(options)
+      options = @options.merge(options)
       packages = Array(packages)
       raise Error, "Multiple packages are not supported yet." if packages.length > 1
 
@@ -298,7 +298,11 @@ module ActiveShipping
           xml.VariableOptions('SATURDAY_DELIVERY')
 
           xml.RequestedShipment do
-            xml.ShipTimestamp(ship_timestamp(options[:turn_around_time]).iso8601(0))
+            if options[:pickup_date]
+              xml.ShipTimestamp(options[:pickup_date].to_time.iso8601(0))
+            else
+              xml.ShipTimestamp(ship_timestamp(options[:turn_around_time]).iso8601(0))
+            end
 
             freight = has_freight?(options)
 
@@ -470,32 +474,46 @@ module ActiveShipping
       message = response_message(xml)
 
       if success
+        missing_xml_field = false
         rate_estimates = xml.root.css('> RateReplyDetails').map do |rated_shipment|
-          service_code = rated_shipment.at('ServiceType').text
-          is_saturday_delivery = rated_shipment.at('AppliedOptions').try(:text) == 'SATURDAY_DELIVERY'
-          service_type = is_saturday_delivery ? "#{service_code}_SATURDAY_DELIVERY" : service_code
+          begin
+            service_code = rated_shipment.at('ServiceType').text
+            is_saturday_delivery = rated_shipment.at('AppliedOptions').try(:text) == 'SATURDAY_DELIVERY'
+            service_type = is_saturday_delivery ? "#{service_code}_SATURDAY_DELIVERY" : service_code
 
-          transit_time = rated_shipment.at('TransitTime').text if ["FEDEX_GROUND", "GROUND_HOME_DELIVERY"].include?(service_code)
-          max_transit_time = rated_shipment.at('MaximumTransitTime').try(:text) if service_code == "FEDEX_GROUND"
+            transit_time = rated_shipment.at('TransitTime').text if ["FEDEX_GROUND", "GROUND_HOME_DELIVERY"].include?(service_code)
+            max_transit_time = rated_shipment.at('MaximumTransitTime').try(:text) if service_code == "FEDEX_GROUND"
 
-          delivery_timestamp = rated_shipment.at('DeliveryTimestamp').try(:text)
+            delivery_timestamp = rated_shipment.at('DeliveryTimestamp').try(:text)
+            delivery_range = delivery_range_from(transit_time, max_transit_time, delivery_timestamp, (service_code == "GROUND_HOME_DELIVERY"), options)
 
-          delivery_range = delivery_range_from(transit_time, max_transit_time, delivery_timestamp, options)
+            currency = rated_shipment.at('RatedShipmentDetails/ShipmentRateDetail/TotalNetCharge/Currency').text
 
-          currency = rated_shipment.at('RatedShipmentDetails/ShipmentRateDetail/TotalNetCharge/Currency').text
-          RateEstimate.new(origin, destination, @@name,
-               self.class.service_name_for_code(service_type),
-               :service_code => service_code,
-               :total_price => rated_shipment.at('RatedShipmentDetails/ShipmentRateDetail/TotalNetCharge/Amount').text.to_f,
-               :currency => currency,
-               :packages => packages,
-               :delivery_range => delivery_range)
+            RateEstimate.new(origin, destination, @@name,
+                 self.class.service_name_for_code(service_type),
+                 :service_code => service_code,
+                 :total_price => rated_shipment.at('RatedShipmentDetails/ShipmentRateDetail/TotalNetCharge/Amount').text.to_f,
+                 :currency => currency,
+                 :packages => packages,
+                 :delivery_range => delivery_range)
+          rescue NoMethodError
+            missing_xml_field = true
+            nil
+          end
         end
+
+        rate_estimates = rate_estimates.compact
+        logger.warn("[FedexParseRateError] Some fields where missing in the response: #{response}") if logger && missing_xml_field
 
         if rate_estimates.empty?
           success = false
-          message = "No shipping rates could be found for the destination address" if message.blank?
+          if missing_xml_field
+            message = "The response from the carrier contained errors and could not be treated"
+          else
+            message = "No shipping rates could be found for the destination address" if message.blank?
+          end
         end
+
       else
         rate_estimates = []
       end
@@ -503,13 +521,15 @@ module ActiveShipping
       RateResponse.new(success, message, Hash.from_xml(response), :rates => rate_estimates, :xml => response, :request => last_request, :log_xml => options[:log_xml])
     end
 
-    def delivery_range_from(transit_time, max_transit_time, delivery_timestamp, options)
+    def delivery_range_from(transit_time, max_transit_time, delivery_timestamp, is_home_delivery, options)
       delivery_range = [delivery_timestamp, delivery_timestamp]
 
       # if there's no delivery timestamp but we do have a transit time, use it
       if delivery_timestamp.blank? && transit_time.present?
         transit_range  = parse_transit_times([transit_time, max_transit_time.presence || transit_time])
-        delivery_range = transit_range.map { |days| business_days_from(ship_date(options[:turn_around_time]), days) }
+        pickup_date = options[:pickup_date] || ship_date(options[:turn_around_time])
+
+        delivery_range = transit_range.map { |days| business_days_from(pickup_date, days, is_home_delivery) }
       end
 
       delivery_range
@@ -528,20 +548,30 @@ module ActiveShipping
       LabelResponse.new(success, message, response_info, {labels: labels})
     end
 
-    def business_days_from(date, days)
+    def business_days_from(date, days, is_home_delivery=false)
       future_date = date
       count       = 0
 
       while count < days
         future_date += 1.day
-        count += 1 if business_day?(future_date)
+        if is_home_delivery
+          count += 1 if home_delivery_business_day?(future_date)
+        else
+          count += 1 if business_day?(future_date)
+        end
       end
 
       future_date
     end
 
+    #Transit times for FedEx® Ground do not include Saturdays, Sundays, or holidays.
     def business_day?(date)
       (1..5).include?(date.wday)
+    end
+
+    #Transit times for FedEx® Home Delivery, do not include Sundays, Mondays, or holidays.
+    def home_delivery_business_day?(date)
+      (2..6).include?(date.wday)
     end
 
     def parse_tracking_response(response, options)
@@ -622,11 +652,12 @@ module ActiveShipping
 
           location = Location.new(:city => city, :state => state, :postal_code => zip_code, :country => country)
           description = event.at('EventDescription').text
+          type_code = event.at('EventType').text
 
           time          = Time.parse(event.at('Timestamp').text)
           zoneless_time = time.utc
 
-          shipment_events << ShipmentEvent.new(description, zoneless_time, location)
+          shipment_events << ShipmentEvent.new(description, zoneless_time, location, description, type_code)
         end
         shipment_events = shipment_events.sort_by(&:time)